requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
psycopg2==2.7.6.1
requests==2.21.0
tqdm==4.28.1
urllib3==1.26.5
xlrd==1.2.0
sqlparse==0.2.4
docopt
PyYAML>=5.1
rollbar==0.15.0
algoliasearch==2.6.1
sshtunnel==0.4.0
